Renamed geoip source/header files due to conflict on Cygwin.
