+++ src/geoip1.c
@@ -37,7 +37,7 @@
 #include <GeoIPCity.h>
 #endif
 
-#include "geoip.h"
+#include "geoip1.h"
 
 #include "error.h"
 #include "util.h"
