+++ src/gholder.c
@@ -40,7 +40,7 @@
 #endif
 
 #ifdef HAVE_GEOLOCATION
-#include "geoip.h"
+#include "geoip1.h"
 #endif
 
 #include "gholder.h"
