+++ src/parser.c
@@ -63,7 +63,7 @@
 #endif
 
 #ifdef HAVE_GEOLOCATION
-#include "geoip.h"
+#include "geoip1.h"
 #endif
 
 #include "parser.h"
