+++ src/tcbtdb.c
@@ -40,7 +40,7 @@
 #include "tcabdb.h"
 
 #ifdef HAVE_GEOLOCATION
-#include "geoip.h"
+#include "geoip1.h"
 #endif
 
 #include "error.h"
